cameraapplication: always assume not-desktop for click packages
Qt platform isn't a good way to determine this. Mainline devices use
Wayland platform and it's expected that Hybris devices will folow suite
soon. We want to detect PinePhone as phone, not desktop, so that the
viewfinder for back camera isn't mirrored.

For now, assume that if we're built for click, we're mobile, as
currently desktop doesn't support Click packages.