qml/Viewfinder: fix content export preview corruption in !hasEXIF mode
Because of signal connection order, content export preview will start
rendering image while `deleteEXIF()` re-writes the file. This causes
preview image to corrupt.

Move all imageSaved() signal handler code from `ViewFinderOverlay` to
`ViewFinderView` so that we can control the ordering. Order it so that
(synchronous) `deleteEXIF()` happens before content export preview
renders image.

While we're at it, remove call to non-existent `fileOperations.
setEXIFData()` (which is redundant with another part of code in
`ViewFinderOverlay`); the exception thrown will prevent the rest of the
code to run.

Finally, add a few comments so that we turn `deleteEXIF()` to be
asynchronous in the future.

Fixes: https://gitlab.com/ubports/development/apps/lomiri-camera-app/-/issues/217
Fixes: e38dbba16647 ("add fixes to barcodereader app and drop obsolete imports")